chore: move interpreter/*_hook.rs to hook/
