+++ src/query/service/src/interpreters/hook/mod.rs
@@ -12,22 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod grant;
-mod metrics;
-mod query_log;
-mod stream;
-mod table;
-mod task;
-mod util;
+mod compact_hook;
+mod refresh_hook;
 
-pub use grant::validate_grant_object_exists;
-pub use query_log::InterpreterQueryLog;
-pub use stream::build_update_stream_meta_seq;
-pub use table::check_referenced_computed_columns;
-pub use task::get_client_config;
-pub use task::make_schedule_options;
-pub use task::make_warehouse_options;
-pub use util::check_deduplicate_label;
-pub use util::create_push_down_filters;
-
-pub use self::metrics::*;
+pub use compact_hook::*;
+pub use refresh_hook::hook_refresh;
+pub use refresh_hook::RefreshDesc;
